move conflateResponses to apollo-runtime
